requirements.txt
Django==1.11.17
djangorestframework==3.6.4
uWSGI==2.0.17.1
# psycopg2==2.8.4
django-suit==0.2.25
django-model-utils==3.2.0
django-extensions==2.1.3
django-extra-fields==2.0.5
django-easy-select2==1.5.5
requests==2.23.0
django-filter==1.0.4